proto: MsgSkippedEventTx eth_*_index as uint64
- to alleviate potential for overflow
- discussed here:
- https://github.com/fuel-infrastructure/fuel-sequencer/pull/466/files#r1977607154

+++ app/abci/app_test.go
@@ -166,8 +166,8 @@ func (s *AppTestSuite) EncodeMsgIndexWithEvents(
 func (s *AppTestSuite) EncodeMsgSkippedEventTx(
 	reasonForSkip string,
 	blockNumber uint64,
-	logIndex uint32,
-	txIndex uint32,
+	logIndex uint64,
+	txIndex uint64,
 	txHash string,
 	sequence uint64,
 ) (tx []byte) {
